Move constraints from install_command to deps
It's preferred not to have constraints in install_command as that
can cause problems with things like lower-constraints jobs which may
end up with conflicting constraints files. install_command is removed
entirely as it is no longer necessary.

Also updates the location of the constraints file to the releases
address as recommended by the release team.

Change-Id: I1415f66454d7394b6cadb5b0c9bb5b9666d10199