Fix doc build in test script
There is some additional setup required for the docs to be built.
We need to have an implementation so there is something to add to
the api docs, and openstackdocstheme requires a git commit so it
can get the last modified timestamp. Also, we no longer generate a
Changelog since we aren't using pbr now so we need to remove that
or we have a missing file.

Change-Id: Ic79aa0fb8f7e5813bec420c8b2690b0e9ef79cc2