Fix cookiecutter tox tests in gate
It looks like for some reason the script is getting a relative path
in ci, which does not work because we do a find on that path at the
end of the script. This is problematic because at that point we've
already cd'd into a different directory so the relative path is no
longer valid.

To fix this, the find is changed to just look at the current
directory, which is the path we wanted to look at anyway.

Change-Id: I13f046d0df2491d3bba04c3feb0fbac7cd188d3c