Use recommended configuration for apidoc
The documentation for apidoc recommends pointing the module_dir at
the module directory itself rather than the top-level directory for
the project so you don't have to explicitly omit setup.py or other
unpackaged Python files.

I did have to remove the _* exclusion because it was breaking in the
current version of Sphinx. It doesn't seem to be necessary in my
testing either. Private files are ignored anyway.

Change-Id: I2f1afcb2a20c3ee0bf8d52c294836e725b5a4e28